removed old sdem examples and updated.
