+++ dem/README.md
@@ -23,7 +23,7 @@ Nevertheless, when using the Swimming DEM application, the following application
 
 Every use case in the DEM Application is composed of a folder containing all the necessary files to run the simulation. Some of the files are used by the GiD GUI and some others by Kratos. The most important ones are the following:
 
-* KratosDEM.py or KratosSwimmingDEM.py: they are the main scripts of the examples. We call them using python to run the simulations. 
+* KratosDEMAnalysis.py or KratosSwimmingDEMAnalysis.py: they are the main scripts of the examples. We call them using python to run the simulations.
 * _example\_name_.mdpa: it contains the information about the _modelpart_. It lists the material properties, nodes, elements, conditions and _submodelparts_.
 * ProjectParametersDEM.json: it includes the main parameters of the example such as time steps, solver types, list of variables to print, etc.
 
@@ -34,12 +34,12 @@ Every use case in the DEM Application is composed of a folder containing all the
 Open a terminal, go to the folder of the case and type:
 
 >
-    python3 KratosDEM.py
+    python3 KratosDEMAnalysis.py
 
 or
 
 >
-    python3 KratosSwimmingDEM.py
+    python3 KratosSwimmingDEMAnalysis.py
 
 
 *Windows*
@@ -48,13 +48,13 @@ First create a "run.bat" file inside the example folder. It should contain the p
 
 > set PATH=C:\\KratosInstall;C:\\KratosInstall\\libs;%PATH%
 >
-> C:\\KratosInstall\\runkratos KratosDEM.py
+> C:\\KratosInstall\\runkratos KratosDEMAnalysis.py
 
 or
 
 > set PATH=C:\\KratosInstall;C:\\KratosInstall\\libs;%PATH%
 >
-> C:\\KratosInstall\\runkratos KratosSwimmingDEM.py
+> C:\\KratosInstall\\runkratos KratosSwimmingDEMAnalysis.py
 
 After that, open a Windows command line, go to the problem folder launch the "run.bat" file to run the simulation:
