--- dem/use_cases/D_DEMPack2_Tutorial_4.gid/D_DEMPack2_Tutorial_4.tree
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<gid>
-  <pre>
-    <groups_open>
-    </groups_open>
-    <layers_open>
-    </layers_open>
-  </pre>
-</gid>
