--- dem/use_cases/D_DEMPack2_Tutorial_4.gid/D_DEMPack2_Tutorial_4.vv
@@ -1,37 +1,0 @@
-BeginZE /home/farrufat/Downloads/D_DEMPack2_Tutorial_4.gid/D_DEMPack2_Tutorial_4.vv
-   x -0.40001264214515686 0.598650157451629639
-   y 0.539785802364349365 -0.254346400499343872
-   z -2.12995624542236328 0.717766642570495605
-   e 0.699999988079071045
-   v 0.537069302639248125 -0.0543155737781175832 -0.204245525246379317
-   r 0.816633479997188605
-   m 0.910667181015014648 0.275245010852813721 -0.308099836111068726 0 -0.391929924488067627 0.339680671691894531 -0.854990124702453613 0 -0.130676165223121643 0.899364948272705078 0.417212754487991333 0 0 0 0 1
-   c 0.159390196204185486 0.000393172813346609473 0.706094861030578613
-   pd 0
-   pno 0
-   pfo 0
-   pf 4
-   pv 0
-   NowUse 1
-   DrawingType 1
-   LightVector -10.15625 -44.5414810180664062 141.881134033203125 0
-   OGLShadow(UseShadows) 0
-   OGLShadow(RenderMode) RenderNice
-   OGLShadow(ShadowBias) -0.002
-   OGLShadow(AmbientShadows) 1
-   OGLShadow(UseFBO) 0
-   OGLShadow(FBOSize) 8192
-   OGLShadow(DimmedLightFactor) 0.02
-   OGLStereo(UseStereo) 0
-   OGLStereo(RelativeEyeDistance) 1
-   OGLStereo(FlipEyes) 0
-   OGLStereo(StereoMode) AnaglyphColors
-   OGLStereo(EyeDistance) 0.075
-   OGLStereo(LeftEyeDistance) 1 0 0 1
-   OGLStereo(RightEyeDistance) 0 1 1 1
-EndZE /home/farrufat/Downloads/D_DEMPack2_Tutorial_4.gid/D_DEMPack2_Tutorial_4.vv
-
-BeginMirror
-   Mirror 0 Angle 0.0 AngleFake -15.0 SizeFactor 2.0 HeightOffset 0.25 Color #d9e5f24d Backlight Off EdgeStyle No Type Scene DrawMode Mirror
-EndMirror
-
