--- dem/use_cases/D_DEMPack2_Tutorial_4.gid/D_DEMPack2_Tutorial_4DEM_Inlet.mdpa
@@ -1,86 +1,0 @@
-Begin ModelPartData
-// VARIABLE_NAME value
-End ModelPartData
-
-Begin Properties 2 // Inlet group: DEM_inlet
-  PARTICLE_DENSITY 1500
-  YOUNG_MODULUS 1.0e5
-  POISSON_RATIO 0.20
-  FRICTION 0.9999999999999999
-  PARTICLE_COHESION 0.0
-  COEFFICIENT_OF_RESTITUTION 0.1
-  PARTICLE_MATERIAL Hertz
-  ROLLING_FRICTION 0.01
-  ROLLING_FRICTION_WITH_WALLS 0.5
-  PARTICLE_SPHERICITY 1.0
-  DEM_DISCONTINUUM_CONSTITUTIVE_LAW_NAME DEM_D_Linear_viscous_Coulomb
-  DEM_CONTINUUM_CONSTITUTIVE_LAW_NAME DEMContinuumConstitutiveLaw
-End Properties
-
-Begin Nodes // GUI Inlet group identifier: DEM_inlet
-12777 0.0108821 0.0181462 0.100000
-12782 0.0111603 0.0262500 0.100000
-12784 0.0108821 0.0343538 0.100000
-12789 0.0167500 0.0111603 0.100000
-12802 0.0196684 0.0198419 0.100000
-12814 0.0167500 0.0413397 0.100000
-12823 0.0215000 0.0326795 0.100000
-12827 0.0262500 0.0111603 0.100000
-12846 0.0262500 0.0413397 0.100000
-12847 0.0314681 0.0196551 0.100000
-12859 0.0357500 0.0111603 0.100000
-12860 0.0326682 0.0329435 0.100000
-12869 0.0341116 0.0419869 0.100000
-12879 0.0417116 0.0181131 0.100000
-12884 0.0413397 0.0262500 0.100000
-12889 0.0413397 0.0357500 0.100000
-End Nodes
-
-Begin SubModelPart 2 // Group name: DEM_inlet
-  Begin SubModelPartData // DEM-Inlet. Group name: DEM_inlet
-  PROPERTIES_ID 2
-  RIGID_BODY_MOTION 0
-  TABLE_NUMBER 0
-  //TABLE_VELOCITY_COMPONENT 0
-  IDENTIFIER DEM_inlet
-  INJECTOR_ELEMENT_TYPE SphericParticle3D
-  ELEMENT_TYPE SphericSwimmingParticle3D
-  CONTAINS_CLUSTERS 0
-  VELOCITY [3] (0.0, 0.0, 1.0)
-  MAX_RAND_DEVIATION_ANGLE 1
-  INLET_NUMBER_OF_PARTICLES 200.0
-  IMPOSED_MASS_FLOW_OPTION 0
-  MASS_FLOW 0.5
-  INLET_START_TIME 0.0
-  INLET_STOP_TIME 1000.0
-  RADIUS 0.0005
-  PROBABILITY_DISTRIBUTION normal
-  STANDARD_DEVIATION 0.0
-  RANDOM_ORIENTATION 1
-  ORIENTATION [4] (0.0, 0.0, 0.0, 0.0)
-  End SubModelPartData
-  Begin SubModelPartNodes
-  12777
-  12782
-  12784
-  12789
-  12802
-  12814
-  12823
-  12827
-  12846
-  12847
-  12859
-  12860
-  12869
-  12879
-  12884
-  12889
-  End SubModelPartNodes
-End SubModelPart
-
-Begin Table 0 TIME VELOCITY
-0.0  0.0
-1.0  0.0
-End Table
-
