--- dem/use_cases/D_DEMPack2_Tutorial_4.gid/define_output.py
@@ -1,12 +1,0 @@
-
-def DefineOutputPoints():
-    output_points = []
-    return output_points 
-
-def DefineCutPlanes():
-    cut_planes_list = []
-    return cut_planes_list
-
-def DefineDragList():
-    drag_list = []
-    return drag_list
