+++ dem/use_cases/bucket/non_uniform_2.gid/KratosDEMAnalysis.py
@@ -11,12 +11,12 @@ solution.Run() """
 class DEMAnalysisStageWithFlush(DEMAnalysisStage):
 
     def __init__(self, model, project_parameters, flush_frequency=10.0):
-        super(DEMAnalysisStageWithFlush,self).__init__(model, project_parameters)
+        super().__init__(model, project_parameters)
         self.flush_frequency = flush_frequency
         self.last_flush = time.time()
 
     def FinalizeSolutionStep(self):
-        super(DEMAnalysisStageWithFlush,self).FinalizeSolutionStep()
+        super().FinalizeSolutionStep()
 
         if self.parallel_type == "OpenMP":
             now = time.time()
@@ -32,4 +32,3 @@ if __name__ == "__main__":
 
     model = KratosMultiphysics.Model()
     DEMAnalysisStageWithFlush(model, project_parameters).Run()
-
