+++ dem/use_cases/bucket/uniform_3.gid/KratosDEMAnalysis.py
@@ -11,12 +11,12 @@ solution.Run() """
 class DEMAnalysisStageWithFlush(DEMAnalysisStage):
 
     def __init__(self, model, project_parameters, flush_frequency=10.0):
-        super(DEMAnalysisStageWithFlush,self).__init__(model, project_parameters)
+        super().__init__(model, project_parameters)
         self.flush_frequency = flush_frequency
         self.last_flush = time.time()
 
     def FinalizeSolutionStep(self):
-        super(DEMAnalysisStageWithFlush,self).FinalizeSolutionStep()
+        super().FinalizeSolutionStep()
 
         if self.parallel_type == "OpenMP":
             now = time.time()
