--- dem/use_cases/swimming_dem.gid/.directory
@@ -1,4 +1,0 @@
-[Dolphin]
-Timestamp=2018,2,19,14,49,43
-Version=4
-ViewMode=2
