--- dem/use_cases/swimming_dem.gid/README.md
@@ -1,18 +1,0 @@
-# Swimming DEM example
-
-**Author:** Salva Latorre
-
-## Case Specification
-
-This case simulates the dragging of two different groups of particles under the action of a flow inside a tube. This example includes an inlet of particles. As already said, this problem includes the interaction of DEM elements with a FEM fluid. The user can study the behaviour of the DEM particles when using different types of fluids (nentonian or non-newtonian), different drag, flow velocities, etc. There is a huge number of postprocess variables available to be printed, which can provide the user with a lot of information.
-
-[<img
-  src="data/swimming_dem1.png?raw=true"
-  width="400"
-  title="Swimming DEM example. Starting point.">
-](data/swimming_dem1.png)
-[<img
-  src="data/swimming_dem2.png?raw=true"
-  width="400"
-  title="Swimming DEM example. Particles rising.">
-](data/swimming_dem2.png)
