--- dem/use_cases/swimming_dem.gid/swimming_dem.model
@@ -1,18 +1,0 @@
-<class name='SphereMeshVariables'>
-<variables owner='MODEL'>
-MinRadiusFactor 0.5
-MaxRadiusFactor 2
-PartSizeDist 0
-DistributionParam,StandarDeviation 0.1
-DistributionParam,Scale 1.0
-DistributionParam,Width 1.0
-Random 0
-BoundaryTolerance 1.0
-ParticleType 1
-Seed 1512403416
-Advanced 1
-Priority 1
-HighPorosity 0
-Porosity 0.95
-</variables>
-</class>
