--- dem/use_cases/swimming_dem.gid/swimming_dem.tree
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<gid>
-  <pre>
-    <groups_open>
-    </groups_open>
-    <layers_open>
-    </layers_open>
-  </pre>
-</gid>
