--- dem/use_cases/swimming_dem.gid/swimming_demDEM_Inlet.mdpa
@@ -1,119 +1,0 @@
-Begin ModelPartData
-// VARIABLE_NAME value
-End ModelPartData
-
-Begin Properties 2 // Inlet group: DEM_inlet
-  PARTICLE_DENSITY 1500.0
-  YOUNG_MODULUS 1.0e5
-  POISSON_RATIO 0.20
-  FRICTION 0.9999999999999999
-  PARTICLE_COHESION 0.0
-  COEFFICIENT_OF_RESTITUTION 0.2
-  PARTICLE_MATERIAL Hertz
-  ROLLING_FRICTION 0.01
-  ROLLING_FRICTION_WITH_WALLS 0.05
-  PARTICLE_SPHERICITY 1.0
-  DEM_DISCONTINUUM_CONSTITUTIVE_LAW_NAME DEM_D_Linear_viscous_Coulomb
-  DEM_CONTINUUM_CONSTITUTIVE_LAW_NAME DEMContinuumConstitutiveLaw
-End Properties
-
-Begin Nodes // GUI Inlet group identifier: DEM_inlet
-16845 0.00250000 0.00250000 0.100000
-16851 0.00250000 0.0120000 0.100000
-16866 0.00250000 0.0215000 0.100000
-16890 0.00250000 0.0310000 0.100000
-16927 0.00250000 0.0405000 0.100000
-16948 0.0120000 0.00250000 0.100000
-16952 0.0113591 0.0154406 0.100000
-16969 0.00250000 0.0500000 0.100000
-16973 0.0111706 0.0262500 0.100000
-17005 0.0113591 0.0370594 0.100000
-17028 0.0184841 0.0105523 0.100000
-17065 0.0215000 0.00250000 0.100000
-17079 0.0120000 0.0500000 0.100000
-17089 0.0215000 0.0198411 0.100000
-17104 0.0202371 0.0329046 0.100000
-17112 0.0184841 0.0419477 0.100000
-17137 0.0271266 0.0109963 0.100000
-17171 0.0310000 0.00250000 0.100000
-17177 0.0215000 0.0500000 0.100000
-17202 0.0262500 0.0413294 0.100000
-17221 0.0331840 0.0195974 0.100000
-17226 0.0311682 0.0310412 0.100000
-17274 0.0370917 0.0109867 0.100000
-17291 0.0310000 0.0500000 0.100000
-17298 0.0405000 0.00250000 0.100000
-17299 0.0343477 0.0416159 0.100000
-17329 0.0416159 0.0181523 0.100000
-17345 0.0413294 0.0262500 0.100000
-17376 0.0413294 0.0357500 0.100000
-17425 0.0405000 0.0500000 0.100000
-17433 0.0500000 0.00250000 0.100000
-17440 0.0500000 0.0120000 0.100000
-17451 0.0500000 0.0215000 0.100000
-17478 0.0500000 0.0310000 0.100000
-17513 0.0500000 0.0405000 0.100000
-17562 0.0500000 0.0500000 0.100000
-End Nodes
-
-Begin SubModelPart 2 // Group name: DEM_inlet
-  Begin SubModelPartData // DEM-Inlet. Group name: DEM_inlet
-  PROPERTIES_ID 2
-  RIGID_BODY_MOTION 0
-  IDENTIFIER DEM_inlet
-  INJECTOR_ELEMENT_TYPE SphericParticle3D
-  ELEMENT_TYPE SphericSwimmingParticle3D
-  CONTAINS_CLUSTERS 0
-  VELOCITY [3] (0.0, 0.0, 1.0)
-  MAX_RAND_DEVIATION_ANGLE 1
-  INLET_NUMBER_OF_PARTICLES 500.0
-  IMPOSED_MASS_FLOW_OPTION 0
-  MASS_FLOW 0.5
-  INLET_START_TIME 0.0
-  INLET_STOP_TIME 1000.0
-  RADIUS 0.0005
-  PROBABILITY_DISTRIBUTION normal
-  STANDARD_DEVIATION 0.0
-  RANDOM_ORIENTATION 1
-  ORIENTATION [4] (0.0, 0.0, 0.0, 0.0)
-  End SubModelPartData
-  Begin SubModelPartNodes
-  16845
-  16851
-  16866
-  16890
-  16927
-  16948
-  16952
-  16969
-  16973
-  17005
-  17028
-  17065
-  17079
-  17089
-  17104
-  17112
-  17137
-  17171
-  17177
-  17202
-  17221
-  17226
-  17274
-  17291
-  17298
-  17299
-  17329
-  17345
-  17376
-  17425
-  17433
-  17440
-  17451
-  17478
-  17513
-  17562
-  End SubModelPartNodes
-End SubModelPart
-
