+++ dem/use_cases/swimming_dem_updated.gid/FluidMaterials.json
@@ -1,0 +1,16 @@
+{
+    "properties" : [{
+        "model_part_name" : "FluidModelPart.FluidParts_fluid",
+        "properties_id"   : 1,
+        "Material"        : {
+            "constitutive_law" : {
+                "name" : "Newtonian3DLaw"
+            },
+            "Variables"        : {
+                "DENSITY"           : 1.225,
+                "DYNAMIC_VISCOSITY" : 1.846e-5
+            },
+            "Tables"           : {}
+        }
+    }]
+}
