+++ dem/use_cases/swimming_dem_updated.gid/swimming_dem_updated.cnd
@@ -1,0 +1,35 @@
+NUMBER: 1 CONDITION: point_Local_axes
+CONDTYPE: over points
+CONDMESHTYPE: over nodes
+QUESTION: Local_Axes#LA#(-GLOBAL-)
+VALUE: -GLOBAL-
+END CONDITION
+NUMBER: 2 CONDITION: line_Local_axes
+CONDTYPE: over lines
+QUESTION: Local_Axes#LA#(-GLOBAL-,-Automatic-,-Automatic_alt-)
+VALUE: -GLOBAL-
+END CONDITION
+NUMBER: 3 CONDITION: surface_Local_axes
+CONDTYPE: over surfaces
+QUESTION: Local_Axes#LA#(-GLOBAL-,-Automatic-,-Automatic_alt-)
+VALUE: -GLOBAL-
+END CONDITION
+NUMBER: 4 CONDITION: volume_Local_axes
+CONDTYPE: over volumes
+CONDMESHTYPE: over body elements
+QUESTION: Local_Axes#LA#(-GLOBAL-)
+VALUE: -GLOBAL-
+END CONDITION
+NUMBER: 5 CONDITION: element_pairs
+CONDTYPE: over volumes
+CONDMESHTYPE: over body elements
+CANREPEAT: yes
+QUESTION: pair_name
+VALUE: -
+END CONDITION
+NUMBER: 6 CONDITION: relation_line_geo_mesh
+CONDTYPE: over lines
+CONDMESHTYPE: over body elements
+QUESTION: id#FUNC#(NumEntity)
+VALUE: 0
+END CONDITION
