+++ dem/use_cases/swimming_dem_updated.gid/swimming_dem_updated.tree
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<gid>
+  <pre>
+    <groups_open>
+    </groups_open>
+    <layers_open>
+    </layers_open>
+  </pre>
+</gid>
