+++ dem/use_cases/swimming_dem_updated.gid/swimming_dem_updatedDEM.mdpa
@@ -1,0 +1,286 @@
+Begin ModelPartData
+//  VARIABLE_NAME value
+End ModelPartData
+
+Begin Properties 2
+    DEM_DISCONTINUUM_CONSTITUTIVE_LAW_NAME DEM_D_Hertz_viscous_Coulomb
+    PARTICLE_DENSITY 2500.0
+    YOUNG_MODULUS 100000.0
+    POISSON_RATIO 0.20
+    FRICTION 0.577350269189494
+    COEFFICIENT_OF_RESTITUTION 0.2
+    PARTICLE_MATERIAL 1
+    ROLLING_FRICTION 0.01
+    ROLLING_FRICTION_WITH_WALLS 0.01
+    PARTICLE_SPHERICITY 1.0
+    DEM_CONTINUUM_CONSTITUTIVE_LAW_NAME DEMContinuumConstitutiveLaw
+End Properties
+
+Begin Nodes
+  542   0.3603354597   0.0034422994   0.4107845769
+  546   0.3724275046   0.0147441468   0.4233666995
+  552   0.3767880110  -0.0094533876   0.4232920377
+  564   0.3689822294   0.0267312852   0.4063389725
+  574   0.3971730388  -0.0001970009   0.4384234608
+  575   0.3603503168   0.0073364648   0.3921164967
+  578   0.3664455500  -0.0137585049   0.3971772892
+  580   0.3848238440   0.0295087580   0.4187278032
+  586   0.3978607654  -0.0243009763   0.4297508864
+  590   0.3782153238   0.0103556549   0.4023054247
+  591   0.3831270543  -0.0309996710   0.4088294039
+  601   0.4036892188   0.0230695474   0.4310376775
+  604   0.3704591118   0.0230449783   0.3863733566
+  609   0.3971301353   0.0083649282   0.4165636735
+  617   0.3856712416   0.0341102318   0.3965033215
+  621   0.3897189086  -0.0082605819   0.3990514321
+  622   0.4014724292  -0.0150569936   0.4134392128
+  624   0.4162217372  -0.0116948228   0.4323241035
+  625   0.4173000515   0.0086331883   0.4330635298
+  643   0.3778386974   0.0008937523   0.3745468001
+  646   0.4069440112   0.0333992675   0.4094564557
+  650   0.3871349878  -0.0270635303   0.3819031364
+  653   0.4172397436  -0.0049373036   0.4158093707
+  657   0.4190757412  -0.0265470640   0.4174003388
+  658   0.4056073935  -0.0362999563   0.4005349187
+  659   0.4083293775   0.0005294019   0.4017696572
+  661   0.4179601449   0.0150732643   0.4131134314
+  665   0.4029704355   0.0221146752   0.3918150687
+  674   0.3901588003   0.0264997823   0.3746619891
+  677   0.4309491977  -0.0145360909   0.4244049292
+  684   0.4161266171  -0.0157555469   0.3993861977
+  685   0.4050191708   0.0378043640   0.3868334292
+  687   0.4326925527   0.0021019415   0.4209346325
+  688   0.4272384547   0.0301887536   0.4128451176
+  696   0.4192676206   0.0143640004   0.3945524545
+  697   0.4253040125  -0.0000148580   0.4016549888
+  699   0.4226175566   0.0307622260   0.3953658346
+  702   0.4154431185  -0.0020915233   0.3832910112
+  707   0.4267947701  -0.0300238651   0.3978251012
+  710   0.4353386918  -0.0139551054   0.4073553737
+  712   0.4020654803  -0.0118843071   0.3646770391
+  713   0.4144428710  -0.0262152107   0.3799890272
+  714   0.3992181045   0.0107115361   0.3613328654
+  716   0.4353325630   0.0166773769   0.4051034621
+  727   0.4151651819   0.0216212974   0.3747539953
+  741   0.4378531007   0.0030232085   0.3924243926
+  742   0.4329012171  -0.0139957181   0.3860912823
+  745   0.4337874174   0.0201069309   0.3864069245
+  748   0.4175951635   0.0032777497   0.3631191783
+  759   0.4313392065   0.0063943708   0.3763934647
+End Nodes
+
+
+Begin Elements SphericSwimmingParticle3D// GUI group identifier: DEM
+     4263          2   741 
+     4264          2   714 
+     4265          2   542 
+     4266          2   748 
+     4267          2   575 
+     4268          2   574 
+     4269          2   687 
+     4270          2   643 
+     4271          2   759 
+     4272          2   580 
+     4273          2   745 
+     4274          2   625 
+     4275          2   546 
+     4276          2   712 
+     4277          2   710 
+     4278          2   586 
+     4279          2   699 
+     4280          2   696 
+     4281          2   727 
+     4282          2   665 
+     4283          2   646 
+     4284          2   685 
+     4285          2   617 
+     4286          2   659 
+     4287          2   702 
+     4288          2   697 
+     4289          2   661 
+     4290          2   653 
+     4291          2   716 
+     4292          2   622 
+     4293          2   621 
+     4294          2   684 
+     4295          2   591 
+     4296          2   658 
+     4297          2   650 
+     4298          2   713 
+     4299          2   742 
+     4300          2   707 
+     4301          2   688 
+     4302          2   657 
+     4303          2   624 
+     4304          2   674 
+     4305          2   604 
+     4306          2   564 
+     4307          2   601 
+     4308          2   677 
+     4309          2   552 
+     4310          2   609 
+     4311          2   590 
+     4312          2   578 
+End Elements
+
+Begin NodalData RADIUS // GUI group identifier: DEM
+        741 0 0.00873772
+        714 0  0.0089291
+        542 0 0.00838781
+        748 0 0.00797554
+        575 0 0.00819707
+        574 0   0.010838
+        687 0 0.00949341
+        643 0  0.0153416
+        759 0 0.00865982
+        580 0  0.0102632
+        745 0  0.0076699
+        625 0 0.00988685
+        546 0 0.00956562
+        712 0  0.0118273
+        710 0  0.0101297
+        586 0  0.0100462
+        699 0 0.00875375
+        696 0  0.0080029
+        727 0  0.0117835
+        665 0 0.00793757
+        646 0  0.0120065
+        685 0 0.00865099
+        617 0  0.0115699
+        659 0 0.00794451
+        702 0  0.0105141
+        697 0 0.00713406
+        661 0 0.00939773
+        653 0 0.00821761
+        716 0  0.0088205
+        622 0 0.00904738
+        621 0   0.010174
+        684 0 0.00913196
+        591 0  0.0112638
+        658 0   0.012012
+        650 0  0.0130061
+        713 0  0.0116891
+        742 0 0.00983994
+        707 0 0.00853897
+        688 0 0.00612396
+        657 0  0.0108674
+        624 0 0.00965518
+        674 0  0.0110956
+        604 0 0.00794498
+        564 0  0.0073607
+        601 0  0.0098881
+        677 0  0.0065728
+        552 0  0.0144028
+        609 0  0.0117509
+        590 0  0.0114503
+        578 0  0.0127873
+End NodalData
+
+Begin SubModelPart DEMParts_DEM // Group DEM // Subtree DEMParts
+    Begin SubModelPartNodes
+          542
+          546
+          552
+          564
+          574
+          575
+          578
+          580
+          586
+          590
+          591
+          601
+          604
+          609
+          617
+          621
+          622
+          624
+          625
+          643
+          646
+          650
+          653
+          657
+          658
+          659
+          661
+          665
+          674
+          677
+          684
+          685
+          687
+          688
+          696
+          697
+          699
+          702
+          707
+          710
+          712
+          713
+          714
+          716
+          727
+          741
+          742
+          745
+          748
+          759
+    End SubModelPartNodes
+    Begin SubModelPartElements
+         4263
+         4264
+         4265
+         4266
+         4267
+         4268
+         4269
+         4270
+         4271
+         4272
+         4273
+         4274
+         4275
+         4276
+         4277
+         4278
+         4279
+         4280
+         4281
+         4282
+         4283
+         4284
+         4285
+         4286
+         4287
+         4288
+         4289
+         4290
+         4291
+         4292
+         4293
+         4294
+         4295
+         4296
+         4297
+         4298
+         4299
+         4300
+         4301
+         4302
+         4303
+         4304
+         4305
+         4306
+         4307
+         4308
+         4309
+         4310
+         4311
+         4312
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    End SubModelPartConditions
+End SubModelPart
