+++ dem/use_cases/swimming_dem_updated.gid/swimming_dem_updatedDEM_FEM_boundary.mdpa
@@ -1,0 +1,6 @@
+Begin ModelPartData
+//  VARIABLE_NAME value
+End ModelPartData
+
+
+
