+++ dem/use_cases/swimming_dem_updated.gid/swimming_dem_updatedDEM_Inlet.mdpa
@@ -1,0 +1,122 @@
+Begin ModelPartData
+//  VARIABLE_NAME value
+End ModelPartData
+
+Begin Properties 3 // Inlet group: DEM_inlet
+    PARTICLE_DENSITY 2500.0
+    YOUNG_MODULUS 100000.0
+    POISSON_RATIO 0.20
+    FRICTION 0.577350269189494
+    COEFFICIENT_OF_RESTITUTION 0.2
+    PARTICLE_MATERIAL 1
+    ROLLING_FRICTION 0.01
+    ROLLING_FRICTION_WITH_WALLS 0.01
+    PARTICLE_SPHERICITY 1.0
+    DEM_DISCONTINUUM_CONSTITUTIVE_LAW_NAME DEM_D_Hertz_viscous_Coulomb
+    DEM_CONTINUUM_CONSTITUTIVE_LAW_NAME DEMContinuumConstitutiveLaw
+End Properties
+
+Begin Nodes
+  817   0.0025000000   0.0025000000   0.1000000000
+  818   0.0025000000   0.0120000000   0.1000000000
+  819   0.0025000000   0.0215000000   0.1000000000
+  820   0.0025000000   0.0310000000   0.1000000000
+  824   0.0025000000   0.0405000000   0.1000000000
+  826   0.0120000000   0.0025000000   0.1000000000
+  827   0.0108820508   0.0181461524   0.1000000000
+  830   0.0025000000   0.0500000000   0.1000000000
+  831   0.0111602540   0.0262500000   0.1000000000
+  832   0.0108820508   0.0343538476   0.1000000000
+  833   0.0167500000   0.0111602540   0.1000000000
+  836   0.0196684028   0.0198418707   0.1000000000
+  837   0.0215000000   0.0025000000   0.1000000000
+  838   0.0120000000   0.0500000000   0.1000000000
+  841   0.0167500000   0.0413397460   0.1000000000
+  842   0.0215000000   0.0326794919   0.1000000000
+  843   0.0262500000   0.0111602540   0.1000000000
+  846   0.0310000000   0.0025000000   0.1000000000
+  847   0.0215000000   0.0500000000   0.1000000000
+  848   0.0262500000   0.0413397460   0.1000000000
+  849   0.0314681120   0.0196550720   0.1000000000
+  851   0.0357500000   0.0111602540   0.1000000000
+  852   0.0326681959   0.0329435408   0.1000000000
+  854   0.0310000000   0.0500000000   0.1000000000
+  855   0.0405000000   0.0025000000   0.1000000000
+  856   0.0341115716   0.0419869348   0.1000000000
+  857   0.0417115716   0.0181130652   0.1000000000
+  858   0.0413397460   0.0262500000   0.1000000000
+  860   0.0413397460   0.0357500000   0.1000000000
+  864   0.0405000000   0.0500000000   0.1000000000
+  866   0.0500000000   0.0025000000   0.1000000000
+  867   0.0500000000   0.0120000000   0.1000000000
+  868   0.0500000000   0.0215000000   0.1000000000
+  870   0.0500000000   0.0310000000   0.1000000000
+  871   0.0500000000   0.0405000000   0.1000000000
+  873   0.0500000000   0.0500000000   0.1000000000
+End Nodes
+
+
+Begin SubModelPart Inlet_DEM_inlet // Group DEM_inlet // Subtree Inlet
+    Begin SubModelPartData
+        PROPERTIES_ID 3
+        RIGID_BODY_MOTION 0
+        //TABLE_VELOCITY_COMPONENT 0
+        IDENTIFIER Inlet_DEM_inlet
+        INJECTOR_ELEMENT_TYPE SphericParticle3D
+        ELEMENT_TYPE SphericSwimmingParticle3D
+        CONTAINS_CLUSTERS 0
+        VELOCITY [3] (1.0, 0.0, 0.0)
+        MAX_RAND_DEVIATION_ANGLE 0.01745329251994
+        INLET_NUMBER_OF_PARTICLES 200.0
+        IMPOSED_MASS_FLOW_OPTION 0
+        INLET_START_TIME 0.0
+        INLET_STOP_TIME 100.0
+        RADIUS 0.5
+        PROBABILITY_DISTRIBUTION normal
+        STANDARD_DEVIATION 0.0
+    End SubModelPartData
+    Begin SubModelPartNodes
+       817
+       818
+       819
+       820
+       824
+       826
+       827
+       830
+       831
+       832
+       833
+       836
+       837
+       838
+       841
+       842
+       843
+       846
+       847
+       848
+       849
+       851
+       852
+       854
+       855
+       856
+       857
+       858
+       860
+       864
+       866
+       867
+       868
+       870
+       871
+       873
+    End SubModelPartNodes
+  End SubModelPart
+    
+        Begin Table 0 TIME VELOCITY
+        0.0  0.0
+        1.0  0.0
+        End Table
+
